Move embedded SVG code that has been reviewed to a mixin
